use dgfgridtype.
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@2230 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ fem/pass/test/dgellipt/main.cc
@@ -7,7 +7,7 @@
 #include <dune/common/misc.hh>
 #include <dune/common/timer.hh>
 
-#include <dune/grid/io/file/dgfparser/gridtype.hh>
+#include <dune/grid/io/file/dgfparser/dgfgridtype.hh>
 #include <dune/common/mpihelper.hh>
 #include <dune/common/exceptions.hh>
 
@@ -60,7 +60,7 @@ int main (int argc, char **argv)
 
     std::string paramfile ( paramname );
 
-    typedef DescriptionTraits <float,GridType,ncomp,2> DescrType;
+    typedef DescriptionTraits <double,GridType,ncomp,2> DescrType;
     typedef DescrType :: ModelType ModelType;
     typedef DescrType :: DiscrParamType DiscrParamType;
